rq: add all tokens
